memory palace/memory palace/requirements.txt
flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0